Added user permissions

+++ app/Http/Livewire/Backoffice/Budgets/Delete.php
@@ -3,10 +3,13 @@
 namespace App\Http\Livewire\Backoffice\Budgets;
 
 use App\Models\Budget;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Component;
 
 class Delete extends Component
 {
+    use AuthorizesRequests;
+
     public $budget;
 
     public function mount(Budget $budget)
@@ -16,6 +19,7 @@ class Delete extends Component
 
     public function delete()
     {
+        $this->authorize('delete', $this->budget);
         $this->budget->delete();
         $this->emit('success', __('budgets.success_delete'), sprintf(__('budgets.message.delete'), $this->budget->event_name));
         $this->emit('hideModal');
@@ -24,6 +28,7 @@ class Delete extends Component
 
     public function render()
     {
+        $this->authorize('delete', $this->budget);
         return view('livewire.backoffice.budgets.delete');
     }
 }
